+++ src/main/java/com/umc5th/dayrecord/service/DiaryService/DroneLocationQueryService.java
@@ -6,7 +6,7 @@ import com.umc5th.dayrecord.domain.DroneLocation;
 public interface DroneLocationQueryService {
 
 
-    DroneLocation save_dl(float latitude,float longitude,boolean isStart);
+    DroneLocation save_dl(double latitude, double longitude,boolean isStart);
     DroneLocation get_dl();
